requirements.txt
django
djangorestframework
django-cors-headers
librosa
torch
openai-whisper
pandas
langchain
frontend
tools
google-generativeai
pillow
opencv-python
langchain_openai
langchain-community
pinecone-client
langchain_pinecone
python-dotenv
pymupdf
sentence-transformers
unstructured
python-docx